+++ lib/puppet/provider/service/windows.rb
@@ -1,8 +1,6 @@
 # Windows Service Control Manager (SCM) provider
 
-require 'win32/service' if Puppet.features.microsoft_windows?
-
-Puppet::Type.type(:service).provide :windows do
+Puppet::Type.type(:service).provide :windows, :parent => :service do
 
   desc <<-EOT
     Support for Windows Service Control Manager (SCM). This provider can
@@ -84,11 +82,6 @@ Puppet::Type.type(:service).provide :windows do
     net(:stop, @resource[:name])
   rescue Puppet::ExecutionFailure => detail
     raise Puppet::Error.new("Cannot stop #{@resource[:name]}, error was: #{detail}" )
-  end
-
-  def restart
-    self.stop
-    self.start
   end
 
   def status
